More refactorizations related to GUI module and sign issues: - Simplified code in regist.cpp - Use size_t instead of int - Removed garbage whitespaces
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@312 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ course.cpp
@@ -1080,7 +1080,7 @@ void CCourse::GetSurfaceType (double x, double z, double weights[]) const {
     }
 }
 
-size_t CCourse::GetTerrainIdx (double x, double z, double level) const {
+int CCourse::GetTerrainIdx (double x, double z, double level) const {
     TIndex2 idx0, idx1, idx2;
     double u, v;
     FindBarycentricCoords (x, z, &idx0, &idx1, &idx2, &u, &v);
@@ -1091,7 +1091,7 @@ size_t CCourse::GetTerrainIdx (double x, double z, double level) const {
 		if (terrain [idx0.i + nx*idx0.j] == i) wheight += u;
 		if (terrain [idx1.i + nx*idx1.j] == i) wheight += v;
 		if (terrain [idx2.i + nx*idx2.j] == i) wheight += 1.0 - u - v;
-     	if (wheight > level) return (i);
+		if (wheight > level) return (int)i;
 	}
 	return -1;
 }
